test: LottoMatcherTest의 테스트를 LottoWinningNumbers로 옮김
